Add build script and load files from resource folder
The new gradle script allows to package a jar for the jSMTLIB library and run the unit tests.
As the unit tests require to load a lot of ressource files from various places,
the source code has been modified in a way to contain all these ressource files in the jar
and loading them from the ressource folder during test execution.