Update build.gradle.kts to build two versions of the jar
We build a version that ships the z3-4.8.5 solver and might be used as a standalone jar.
The solver binaries are load from the jar and
`java -ea -cp build/libs/jSMTLIB-0.9.10.4-with-solvers.jar APIExample`
works as expected.

The other version does not ship any solvers. Only the parser and the logics.
It is smaller and thought for settings where only the parser is used.
We are going to ship this jar along with jConstraints